move config code to module. refactor away distinction of libs/deps
